+++ inst/unit_tests/test_Parallel.R
@@ -19,7 +19,7 @@ test_Parallel <- function() {
                                    pvars,dYName,dYTarget,verbose=FALSE,
                                    parallelCluster=cl)
   dTrainCTreatedP <- prepare(treatmentsCP,uci.car.data,pruneSig=c(),
-                             parallelCluster=cl)
+                             parallelCluster=cl, check_for_duplicate_frames=FALSE)
   if(!is.null(cl)) {
     parallel::stopCluster(cl)
     cl <- NULL
@@ -27,7 +27,7 @@ test_Parallel <- function() {
   set.seed(seedVal)
   treatmentsC <- designTreatmentsC(uci.car.data,
                                     pvars,dYName,dYTarget,verbose=FALSE)
-  dTrainCTreated <- prepare(treatmentsC,uci.car.data,pruneSig=c())
+  dTrainCTreated <- prepare(treatmentsC,uci.car.data,pruneSig=c(), check_for_duplicate_frames=FALSE)
   
   RUnit::checkTrue(nrow(dTrainCTreated)==nrow(dTrainCTreatedP))
   RUnit::checkTrue(length(colnames(dTrainCTreated))==length(colnames(dTrainCTreatedP)))